refactor(App): refactor app.tsx and distribute content
