+++ src/config/defaults.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_PROJECT_ROOT = "/var/www/html";
+export const DEFAULT_PHPMYADMIN_URL = "http://localhost/phpmyadmin";
+export const POLL_INTERVAL_MS = 10_000;
+export const LOG_LINE_COUNT = 80;
