+++ src/config/storage.ts
@@ -1,0 +1,11 @@
+export const storageKeys = {
+  projectRoot: "stackpilot.projectRoot",
+  projectName: "stackpilot.projectName",
+  phpMyAdminUrl: "stackpilot.phpMyAdminUrl",
+  distroPreset: "stackpilot.distroPreset",
+  serviceUnits: "stackpilot.serviceUnits",
+  setupComplete: "stackpilot.setupComplete",
+  theme: "stackpilot.theme",
+  windowsMode: "stackpilot.windowsMode",
+  xamppMode: "stackpilot.xamppMode",
+} as const;
